Tailorized "2005-02-18 17:43:11 by logg" Fixes for PETSc 2.2.1, based on suggestions by Garth Wells.
Added two new options to configure:

    --with-petsc=<path>

    --with-petsc-version=<version>

The configure script first uses the path to PETSc specified by
the flag --with-petsc=<path>. It then tries the environment variable
PETSC_DIR, and if neither is given then uses /usr/lib/petsc.

To compile DOLFIN with PETSc 2.2.1 (which is not yet available in Debian),
use the option --with-petsc-version=2.2.1. This should take care of some
of the compiler flags and the changed API for the Krylov solvers.

Please check that it works and tell me if something needs to be changed.