Tailorized "2005-12-01 13:16:10 by garth" Functions for applying boundary conditions to the RHS vector only now accept the solution vector x. Where Dirichlet boundary conditions are applied, b = x - bc.The is useful for applying boundary conditions within Newton iterations.
This is not consistent with the other applyBC functions which return b = bc (no minus sign). This is due to the definition of the nonlinear function F(u)=0.
For Poisson's equation, I've been using the format F(u) = (grad v,grad u) - (v,f), hence the need for the minus. Is there a preference as to how we set-up nonlinear problems?

Would it be better to use a different name for funtions that return the difference between the current approximate solution and the Dirichlet BC?