Tailorized "2005-12-02 16:14:09 by garth" Updates to nls demos.
The nonlinear solver demos solve using both the class NonlinearSolver (which exploits the PETSc Newton solvers) and a hand programmed Newton procedure. The PETSc solvers are typically about 10% faster for the tests I've done. However, the NonlinearSolver class and the interface to it would be simpler if I programmed the NewtonSolver, and just relied on the existing linear algerbra wrappers for PETSc vectors, matrices and linear solvers. I think that this would be better given that the speed-up using the PETSc solver is not so large, we could probably equal it with some with some optimisations. Also, we then restrict PETSc to the linear algerbra classes, limiting the intrusion of PETSc into DOLFIN. Are there reasons to stick with the PETSc Newton solvers that I might be missing?

Garth