Tailorized "2005-12-22 20:24:12 by johanjan" ODE:
Added a conservative modification to the multi-adaptivity which
produces much better results for the reaction benchmark. Some
parameters are hard-coded and chosen ad-hoc, this needs to be fixed in
the future. The multi-adaptivity now also uses the maximum local error
(and the appropriate residual and timestep) over the slab, so there
should no longer be under-estimation of the local error. It's not been
tested yet if this has any significant effect on the reaction
benchmark, Minor updates to the reaction benchmark to conform to the
new parameter system.