Tailorized "2006-01-09 01:44:55 by johanjan" ODE:
Progress on conservative modification for multi-adaptivity:
implemented simple condition which seems to work well generally.

Added results for "domain" version of reaction benchmark. The
multi-adaptive error is now below the mono-adaptive error. The
multi-adaptive gain is very large (a factor 10+, looks like 20 in the
limit), so these results are good.