Clean up of nonlinear solver.
NonlinearFunction is now a simple base class, for which a user provides functions to compute a RHS vector and the Jacobian matrix.

What was NonlinearFunction is now largely NonlinearPDE, which is not dissimilar to the class PDE. Given a BilinearForm, LinearForm, Mesh, and BoundaryCondition, it will form the residual vector and the Jacobian, or this can be provided by the user by deriving a class from NonlinearPDE.

NewtonSolver now uses the new parameters system. Defaults are set in the constructor and have been removed from DefaultParameters.h.