PyDOLFIN:
Major restructure of interface files and compilation. The feature
"-includeall" is now used, which recursively follows all header
files. Selective inclusion is performed by controlling which include
paths are set.

Passes all unit tests, and generally seems to work just as before.

This change should simplify breaking out the modules as a separate
target.

General:

A few (3 I think) minor interface changes to avoid violating the SWIG
parser, "define" cannot be used as a method name because it's a
keyword for example.