Add support for setting timers and printing summaries of tasks.
A timer can be used for timing tasks. The basic usage is

  Timer timer("Assembling over cells");

The timer is started at construction and timing ends
when the timer is destroyed (goes out of scope). It is
also possible to start and stop a timer explicitly by

  timer.start();
  timer.stop();

Timings are stored globally and a summary may be printed
by calling

  summary();

Summary report needs to be improved.