Work on new sub Function logic.
The new logic is more consistent with conventions. Function::operator[](uint i) now returns a reference to the sub Function. A Function keeps track of its sub Functions. Therefore,

    Function& u0 = u[0];

and

   plot(u[0]);

do not involve copying a vector. u and u0 share the same vector. If a DofMap is renumbered after construction, the Function can make sure that all its sub Functions are modified accordingly. Likewise if the mesh is refined.

Assignment makes a deep copy, i.e. you will get a new copy of everything. Therefore with

  Function u0 = u[0];

a new vector will be created and u0 will not change when u changes.

What remains to be done is reseting the dof map for the reduced dimension of the subfunction and extracting vector components upon assignment. Eventually, for

  Function u0 = u[0];
  GenericVector& x = u0.vector();

the vector x will only contains components related to the sub function u0.