Add SUPG stabilisation to adevction-diffusion Python demo.
I had to change the velocity field from the computed Stokes field to a constant vector. Also, the boundary conditions had to be changed to avoid a discontinuity. To use the Stokes field, a shock capturing terms needs to be added since the Stokes problem has some very high gradients.

C++ demo still needs to be updated.