Add Sun Grid Engine (sge) in addition to torque to the available backends for the submit function. Others that uses sge can please have a look at the implementation. I am not that familiar to sge...
To choose sge as backend just add:

  submit(..., backend="sge", parallel_environment="orte", queue="all.d")

The sge backend requires a user to choose a parallel_environment (PE) and 
a queue. The available queues and PE can be aquired by query qconf:

  $> qconf -sql
  $> qconf -spl

respectively. If the queue argument is not set it defaults to "all.d".

The backend defaults to 'torque', and should make the change backward compatible.