Add 'not_working_in_parallel' to XMLMeshFunction.
Hopefully now all demos that are not working in parallel will lead to a 'not working in parallel' error being thrown somewhere in the DOLFIN library.

There are a very small number of cases in which the demo code is clearly not parallel, and these still contain a 'not_working_in_parallel' call.